@@ -1,25 +1,24 @@
-buildscript {
-    repositories {
-        mavenCentral()
-    }
-    dependencies {
-        classpath 'com.android.tools.build:gradle:1.0.1'
-    }
-}
-
 apply plugin: 'com.android.library'
 
-repositories {
-    mavenCentral()
-}
-
 dependencies {
-    compile 'com.android.support:support-v4:+'
-    compile project(':AndroidAsync:AndroidAsync')
-    compile 'com.google.code.gson:gson:2.3'
+    api project(':AndroidAsync:AndroidAsync')
+    api 'com.google.code.gson:gson:2.8.5'
+
+    // Use play services if the linking app is using it. When play services isn't available
+    // ConscryptMiddleware will throw ClassNotFoundErrors upon attempting to load it at runtime.
+    // Conscrypt will fail to load, and the system SSL stack will be used.
+    // "compileOnly" means the library is compiled but not required at runtime.
+    compileOnly 'com.google.android.gms:play-services-base:17.1.0'
+
+    api 'androidx.legacy:legacy-support-v4:1.0.0'
 
     // for when i wanna test this against gms conscrypt
-    androidTestCompile 'com.google.android.gms:play-services:+'
+    androidTestImplementation 'com.google.android.gms:play-services-base:17.1.0'
+
+    // mockito nom nom oh god im drunk
+    androidTestImplementation "org.mockito:mockito-core:"
+    androidTestImplementation "com.google.dexmaker:dexmaker:1.2"
+    androidTestImplementation "com.google.dexmaker:dexmaker-mockito:1.2"
 }
 
 android {
@@ -31,9 +30,9 @@ android {
                 srcDir 'src/'
             }
         }
-        androidTest.java.srcDirs=['test/src/']
-        androidTest.res.srcDirs=['test/res/']
-        androidTest.assets.srcDirs=['test/assets/']
+//        androidTest.java.srcDirs=['test/src/']
+//        androidTest.res.srcDirs=['test/res/']
+//        androidTest.assets.srcDirs=['test/assets/']
     }
 
     lintOptions {
@@ -41,30 +40,17 @@ android {
     }
 
     defaultConfig {
-        targetSdkVersion 21
-        minSdkVersion 9
+        targetSdkVersion 28
+        minSdkVersion 14
     }
 
-    compileSdkVersion 21
-    buildToolsVersion '21.1.2'
-}
-
-// upload to maven task
-if (System.getenv().I_AM_KOUSH == 'true') {
-    apply from: 'https://raw.githubusercontent.com/koush/mvn-repo/master/maven.gradle'
-
-    apply plugin: 'maven'
-    apply plugin: 'signing'
-
-    afterEvaluate { project ->
-        uploadArchives {
-            repositories {
-                mavenDeployer {
-                    pom*.whenConfigured { pom ->
-                        pom.dependencies.removeAll { dep -> dep.artifactId == 'support-v4' }
-                    }
-                }
-            }
-        }
+    compileOptions {
+        sourceCompatibility 1.8
+        targetCompatibility 1.8
     }
+
+    compileSdkVersion project.hasProperty('global_compileSdkVersion') ? global_compileSdkVersion : 25
+    buildToolsVersion project.hasProperty('global_buildToolsVersion') ? global_buildToolsVersion : '25.0.2'
 }
+
+apply from: 'maven.gradle'
