@@ -2,7 +2,6 @@ package com.koushikdutta.ion.sample;
 
 import android.app.Activity;
 import android.content.Context;
-import android.net.Uri;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
@@ -13,6 +12,7 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.GridView;
 import android.widget.ImageView;
+import android.widget.Toast;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
@@ -23,7 +23,7 @@ import com.koushikdutta.ion.Ion;
 /**
  * Created by koush on 6/4/13.
  */
-public class GoogleImageSearch extends Activity {
+public class ImageSearch extends Activity {
     private MyAdapter mAdapter;
 
     // Adapter to populate and imageview from an url contained in the array adapter
@@ -60,9 +60,11 @@ public class GoogleImageSearch extends Activity {
         if (loading != null && !loading.isDone() && !loading.isCancelled())
             return;
 
+        String url = getApiUrl(searchText.getText().toString(), mAdapter.getCount());
+
         // query googles image search api
-        loading = Ion.with(GoogleImageSearch.this)
-        .load(String.format("https://ajax.googleapis.com/ajax/services/search/images?v=1.0&q=%s&start=%d&imgsz=medium", Uri.encode(searchText.getText().toString()), mAdapter.getCount()))
+        loading = Ion.with(ImageSearch.this)
+        .load(url)
         // get the results as json
         .asJsonObject()
         .setCallback(new FutureCallback<JsonObject>() {
@@ -71,19 +73,49 @@ public class GoogleImageSearch extends Activity {
                 try {
                     if (e != null)
                         throw e;
-                    // find the results and populate
-                    JsonArray results = result.getAsJsonObject("responseData").getAsJsonArray("results");
-                    for (int i = 0; i < results.size(); i++) {
-                        mAdapter.add(results.get(i).getAsJsonObject().get("url").getAsString());
-                    }
+                    processApiResult(result);
                 }
                 catch (Exception ex) {
-                    // toast any error we encounter (google image search has an API throttling limit that sometimes gets hit)
-//                    Toast.makeText(GoogleImageSearch.this, ex.toString(), Toast.LENGTH_LONG).show();
+                    // toast any error we encounter (most image search APIs have a throttling limit that sometimes gets hit)
+                    Toast.makeText(ImageSearch.this, ex.toString(), Toast.LENGTH_LONG).show();
                 }
 
             }
         });
+    }
+
+    /**
+     * Build the url for the next request. Uses Flickr's API
+     *
+     * https://www.flickr.com/services/feeds/docs/photos_public/
+     *
+     * @param text Search text
+     * @param page Current page we're in
+     * @return Url for the next request
+     */
+    String getApiUrl(String text, int page) {
+        String base = "https://api.flickr.com/services/feeds/photos_public.gne?format=json&nojsoncallback=?";
+        if (text != null && !text.isEmpty()) {
+            base += "&tags=" + text;
+        }
+
+        return base;
+    }
+
+    /**
+     * Process a successfull API result
+     * @param result the API's response
+     */
+    void processApiResult(JsonObject result) {
+        processFlickrApiResult(result);
+    }
+
+    void processFlickrApiResult(JsonObject result) {
+        // find the results and populate
+        JsonArray results = result.getAsJsonArray("items");
+        for (int i = 0; i < results.size(); i++) {
+            mAdapter.add(results.get(i).getAsJsonObject().getAsJsonObject("media").get("m").getAsString());
+        }
     }
 
     EditText searchText;
@@ -96,7 +128,7 @@ public class GoogleImageSearch extends Activity {
 
         Ion.getDefault(this).configure().setLogging("ion-sample", Log.DEBUG);
 
-        setContentView(R.layout.google_image_search);
+        setContentView(R.layout.image_search);
 
         final Button search = (Button) findViewById(R.id.search);
         searchText = (EditText) findViewById(R.id.search_text);
