@@ -31,6 +31,7 @@ public class IonImageViewRequestBuilder extends IonBitmapRequestBuilder implemen
     ContextReference.ImageViewContextReference imageViewPostRef;
     boolean fadeIn = true;
     boolean crossfade;
+    BitmapDrawableFactory bitmapDrawableFactory = BitmapDrawableFactory.DEFAULT;
 
     public IonImageViewRequestBuilder(IonRequestBuilder builder) {
         super(builder);
@@ -47,6 +48,7 @@ public class IonImageViewRequestBuilder extends IonBitmapRequestBuilder implemen
         crossfade = false;
         imageViewPostRef = null;
         placeholderDrawable = null;
+        bitmapDrawableFactory = BitmapDrawableFactory.DEFAULT;
         placeholderResource = 0;
         errorDrawable = null;
         errorResource = 0;
@@ -83,7 +85,7 @@ public class IonImageViewRequestBuilder extends IonBitmapRequestBuilder implemen
         return this;
     }
 
-    private IonDrawable setIonDrawable(ImageView imageView, BitmapFetcher bitmapFetcher, int loadedFrom) {
+    private IonDrawable setIonDrawable(ImageView imageView, BitmapFetcher bitmapFetcher, ResponseServedFrom servedFrom) {
         BitmapInfo info = null;
         if (bitmapFetcher != null)
             info = bitmapFetcher.info;
@@ -92,13 +94,15 @@ public class IonImageViewRequestBuilder extends IonBitmapRequestBuilder implemen
 
         IonDrawable ret = IonDrawable.getOrCreateIonDrawable(imageView)
         .ion(ion)
-        .setBitmap(info, loadedFrom)
+        .setBitmap(info, servedFrom)
         .setBitmapFetcher(bitmapFetcher)
         .setRepeatAnimation(animateGifMode == AnimateGifMode.ANIMATE)
         .setSize(resizeWidth, resizeHeight)
         .setError(errorResource, errorDrawable)
         .setPlaceholder(placeholderResource, placeholderDrawable)
-        .setFadeIn(fadeIn || crossfade);
+        .setFadeIn(fadeIn || crossfade)
+        .setBitmapDrawableFactory(bitmapDrawableFactory)
+        .updateLayers();
         imageView.setImageDrawable(ret);
         return ret;
     }
@@ -120,13 +124,12 @@ public class IonImageViewRequestBuilder extends IonBitmapRequestBuilder implemen
 
     @Override
     public ImageViewFuture intoImageView(ImageView imageView) {
-        assert Thread.currentThread() == Looper.getMainLooper().getThread();
         if (imageView == null)
             throw new NullPointerException("imageView");
 
         // no uri? just set a placeholder and bail
         if (builder.uri == null) {
-            setIonDrawable(imageView, null, 0).cancel();
+            setIonDrawable(imageView, null, ResponseServedFrom.LOADED_FROM_NETWORK).cancel();
             return ImageViewFutureImpl.FUTURE_IMAGEVIEW_NULL_URI;
         }
 
@@ -162,7 +165,7 @@ public class IonImageViewRequestBuilder extends IonBitmapRequestBuilder implemen
         BitmapFetcher bitmapFetcher = executeCache(sampleWidth, sampleHeight);
         if (bitmapFetcher.info != null) {
             doAnimation(imageView, null, 0);
-            IonDrawable drawable = setIonDrawable(imageView, bitmapFetcher, Loader.LoaderEmitter.LOADED_FROM_MEMORY);
+            IonDrawable drawable = setIonDrawable(imageView, bitmapFetcher, ResponseServedFrom.LOADED_FROM_MEMORY);
             drawable.cancel();
             ImageViewFutureImpl imageViewFuture = ImageViewFutureImpl.getOrCreateImageViewFuture(imageViewPostRef, drawable)
             .setInAnimation(inAnimation, inAnimationResource)
@@ -173,7 +176,7 @@ public class IonImageViewRequestBuilder extends IonBitmapRequestBuilder implemen
             return imageViewFuture;
         }
 
-        IonDrawable drawable = setIonDrawable(imageView, bitmapFetcher, 0);
+        IonDrawable drawable = setIonDrawable(imageView, bitmapFetcher, ResponseServedFrom.LOADED_FROM_NETWORK);
         doAnimation(imageView, loadAnimation, loadAnimationResource);
         ImageViewFutureImpl imageViewFuture = ImageViewFutureImpl.getOrCreateImageViewFuture(imageViewPostRef, drawable)
         .setInAnimation(inAnimation, inAnimationResource)
@@ -269,6 +272,12 @@ public class IonImageViewRequestBuilder extends IonBitmapRequestBuilder implemen
     @Override
     public IonImageViewRequestBuilder animateIn(int animationResource) {
         inAnimationResource = animationResource;
+        return this;
+    }
+
+    @Override
+    public IonImageViewRequestBuilder bitmapDrawableFactory(BitmapDrawableFactory bitmapDrawableFactory) {
+        this.bitmapDrawableFactory = bitmapDrawableFactory;
         return this;
     }
 }
