@@ -6,22 +6,35 @@ import com.koushikdutta.async.http.AsyncHttpRequest;
  * Created by koush on 7/6/13.
  */
 public class Response<T> {
-    AsyncHttpRequest request;
+    public Response(AsyncHttpRequest request, ResponseServedFrom servedFrom, HeadersResponse headers, Exception e, T result) {
+        this.request = request;
+        this.servedFrom = servedFrom;
+        this.headers = headers;
+        this.exception = e;
+        this.result = result;
+    }
+
+    private ResponseServedFrom servedFrom;
+    public ResponseServedFrom getServedFrom() {
+        return servedFrom;
+    }
+
+    private AsyncHttpRequest request;
     public AsyncHttpRequest getRequest() {
         return request;
     }
 
-    T result;
+    private T result;
     public T getResult() {
         return result;
     }
 
-    Exception exception;
+    private Exception exception;
     public Exception getException() {
         return exception;
     }
 
-    HeadersResponse headers;
+    private HeadersResponse headers;
     public HeadersResponse getHeaders() {
         return headers;
     }
