@@ -7,10 +7,10 @@ import android.widget.ProgressBar;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.google.gson.reflect.TypeToken;
+import com.koushikdutta.async.http.NameValuePair;
 import com.koushikdutta.ion.HeadersCallback;
 import com.koushikdutta.ion.ProgressCallback;
 
-import org.apache.http.NameValuePair;
 import org.w3c.dom.Document;
 
 import java.io.File;
