@@ -1,18 +1,21 @@
 package com.koushikdutta.ion.conscrypt;
 
 import android.content.Context;
+import android.util.Log;
 
+import com.google.android.gms.security.ProviderInstaller;
 import com.koushikdutta.async.AsyncSSLSocketWrapper;
 import com.koushikdutta.async.future.Cancellable;
+import com.koushikdutta.async.http.AsyncSSLEngineConfigurator;
 import com.koushikdutta.async.http.AsyncSSLSocketMiddleware;
 import com.koushikdutta.async.http.SimpleMiddleware;
 
-import java.lang.reflect.Method;
 import java.security.Provider;
 import java.security.Security;
 
 import javax.net.ssl.HttpsURLConnection;
 import javax.net.ssl.SSLContext;
+import javax.net.ssl.SSLEngine;
 import javax.net.ssl.SSLSocketFactory;
 
 /**
@@ -24,6 +27,8 @@ public class ConscryptMiddleware extends SimpleMiddleware {
     static boolean success;
     boolean instanceInitialized;
     boolean enabled = true;
+
+    private static final String LOGTAG = "IonConscrypt";
 
     public void enable(boolean enabled) {
         this.enabled = enabled;
@@ -42,55 +47,36 @@ public class ConscryptMiddleware extends SimpleMiddleware {
                 initialized = true;
 
                 // GMS Conscrypt is already initialized, from outside ion. Leave it alone.
-                if (Security.getProvider(GMS_PROVIDER) != null) {
+                if (Security.getProvider(ProviderInstaller.PROVIDER_NAME) != null) {
                     success = true;
                     return;
                 }
 
                 SSLContext originalDefaultContext = SSLContext.getDefault();
                 SSLSocketFactory originalDefaultSSLSocketFactory = HttpsURLConnection.getDefaultSSLSocketFactory();
-                try {
-                    Class<?> providerInstaller = Class.forName("com.google.android.gms.security.ProviderInstaller");
-                    Method mInsertProvider = providerInstaller.getDeclaredMethod("installIfNeeded", Context.class);
-                    mInsertProvider.invoke(null, context);
 
-                } catch (Throwable ignored) {
-                    Context gms = context.createPackageContext("com.google.android.gms", Context.CONTEXT_INCLUDE_CODE | Context.CONTEXT_IGNORE_SECURITY);
-                    gms
-                    .getClassLoader()
-                    .loadClass("com.google.android.gms.common.security.ProviderInstallerImpl")
-                    .getMethod("insertProvider", Context.class)
-                    .invoke(null, context);
-                }
+                ProviderInstaller.installIfNeeded(context);
 
                 Provider[] providers = Security.getProviders();
-                Provider provider = Security.getProvider(GMS_PROVIDER);
-                Security.removeProvider(GMS_PROVIDER);
+                Provider provider = Security.getProvider(ProviderInstaller.PROVIDER_NAME);
+                Security.removeProvider(ProviderInstaller.PROVIDER_NAME);
                 Security.insertProviderAt(provider, providers.length);
                 SSLContext.setDefault(originalDefaultContext);
                 HttpsURLConnection.setDefaultSSLSocketFactory(originalDefaultSSLSocketFactory);
                 success = true;
             }
         }
-        catch (Exception e) {
+        catch (Throwable e) {
+            Log.w(LOGTAG, "Conscrypt initialization failed.", e);
         }
     }
-
-    private static final String GMS_PROVIDER = "GmsCore_OpenSSL";
 
     public void initialize() {
         initialize(context);
         if (success && !instanceInitialized && enabled) {
             instanceInitialized = true;
             try {
-                SSLContext sslContext = null;
-                try {
-                    sslContext = SSLContext.getInstance("TLS", GMS_PROVIDER);
-                }
-                catch (Exception e) {
-                }
-                if (sslContext == null)
-                    sslContext = SSLContext.getInstance("TLS");
+                final SSLContext sslContext = SSLContext.getInstance("TLS", ProviderInstaller.PROVIDER_NAME);
                 sslContext.init(null, null, null);
                 // only set the SSL context if it is the default SSL context
                 if (middleware.getSSLContext() == AsyncSSLSocketWrapper.getDefaultSSLContext())
