@@ -99,7 +99,7 @@ public class CookieMiddleware extends SimpleMiddleware {
     }
 
     @Override
-    public void onHeadersReceived(OnHeadersReceivedDataOnRequestSentData data) {
+    public void onHeadersReceived(OnHeadersReceivedData data) {
         maybeInit();
         try {
             put(URI.create(data.request.getUri().toString()), data.response.headers());
@@ -121,7 +121,7 @@ public class CookieMiddleware extends SimpleMiddleware {
 
             Headers dump = new Headers();
             for (HttpCookie cookie: cookies) {
-                dump.add("Set-Cookie", cookie.getName() + "=" + cookie.getValue());
+                dump.add("Set-Cookie", cookie.getName() + "=" + cookie.getValue() + "; path=" + cookie.getPath());
             }
 
             String key = uri.getScheme() + "://" + uri.getAuthority();
