@@ -13,6 +13,7 @@ import android.provider.MediaStore;
 import com.koushikdutta.async.future.Future;
 import com.koushikdutta.async.future.SimpleFuture;
 import com.koushikdutta.ion.Ion;
+import com.koushikdutta.ion.ResponseServedFrom;
 import com.koushikdutta.ion.bitmap.BitmapInfo;
 
 import java.io.File;
@@ -79,7 +80,7 @@ public class VideoLoader extends SimpleLoader {
                             bmp = Bitmap.createScaledBitmap(bmp, (int) (bmp.getWidth() * ratio), (int) (bmp.getHeight() * ratio), true);
                     }
                     BitmapInfo info = new BitmapInfo(key, type.mimeType, bmp, originalSize);
-                    info.loadedFrom = LoaderEmitter.LOADED_FROM_CACHE;
+                    info.servedFrom = ResponseServedFrom.LOADED_FROM_CACHE;
                     ret.setComplete(info);
                 }
                 catch (OutOfMemoryError e) {
